+++ playground/TestProject.Client.Wasm/_Imports.razor
@@ -12,7 +12,6 @@
 @using Sienar.Identity
 @using Sienar.Infrastructure
 @using Sienar.Layouts
-@using Sienar.Services
 @using Sienar.Ui
 @using TestProject.Client
 @using TestProject.Client.Layouts
