+++ plugins/identity/src/Sienar.Plugins.Identity.Client/Extensions/MenuProviderExtensions.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Components;
 using MudBlazor;
+using Sienar.Data;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
 using Sienar.Menus;
-using Sienar.Services;
 
 namespace Sienar.Extensions;
