+++ plugins/identity/src/Sienar.Plugins.Identity.Client/_Imports.razor
@@ -13,7 +13,6 @@
 @using Sienar.Identity.Requests
 @using Sienar.Infrastructure
 @using Sienar.Data
-@using Sienar.Services
 @using Sienar.Layouts
 @using Sienar.Pages
 @using Sienar.Security
