+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/AccountController.cs
@@ -9,7 +9,6 @@ using Sienar.Data;
 using Sienar.Identity.Requests;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Services;
 
 namespace Sienar.Identity;
