+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/AccountLockoutRequest.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Sienar.Services;
+using Sienar.Data;
 
 namespace Sienar.Identity.Requests;
