+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/AddUserToRoleRequest.cs
@@ -1,4 +1,4 @@
-﻿using Sienar.Services;
+﻿using Sienar.Data;
 
 namespace Sienar.Identity.Requests;
