+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/ChangePasswordRequest.cs
@@ -2,7 +2,7 @@
 
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using Sienar.Services;
+using Sienar.Data;
 using Sienar.Validation;
 
 namespace Sienar.Identity.Requests;
