+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/DeleteAccountRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using Sienar.Services;
+using Sienar.Data;
 
 namespace Sienar.Identity.Requests;
