+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/LockUserAccountRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using Sienar.Services;
+using Sienar.Data;
 
 namespace Sienar.Identity.Requests;
