+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/LoginRequest.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
+using Sienar.Data;
 using Sienar.Security;
-using Sienar.Services;
 
 namespace Sienar.Identity.Requests;
