+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/PerformEmailChangeRequest.cs
@@ -1,6 +1,6 @@
 using System;
 using System.ComponentModel.DataAnnotations;
-using Sienar.Services;
+using Sienar.Data;
 
 namespace Sienar.Identity.Requests;
