+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Requests/ResetPasswordRequest.cs
@@ -1,9 +1,9 @@
 using System;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
+using Sienar.Data;
 using Sienar.Validation;
 using Sienar.Security;
-using Sienar.Services;
 
 namespace Sienar.Identity.Requests;
