+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Results/AccountDataResult.cs
@@ -1,4 +1,4 @@
-﻿using Sienar.Services;
+﻿using Sienar.Data;
 
 namespace Sienar.Identity.Results;
