+++ plugins/identity/src/Sienar.Plugins.Identity.Shared/Identity/Results/PersonalDataResult.cs
@@ -1,5 +1,5 @@
-﻿using Sienar.Infrastructure;
-using Sienar.Services;
+﻿using Sienar.Data;
+using Sienar.Infrastructure;
 
 namespace Sienar.Identity.Results;
