+++ shared/src/Sienar.Utils/Data/IRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Sienar.Services;
+﻿namespace Sienar.Data;
 
 /// <summary>
 /// A marker interface that indicates a type can be used as input to an <c>IGeneralActor</c> or <c>IRequestService</c>
