+++ shared/src/Sienar.Utils/Data/IResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Sienar.Services;
+﻿namespace Sienar.Data;
 
 /// <summary>
 /// A marker interface that indicates a type can be used as the output of an <c>IGeneralActor</c> or <c>IResultActor</c>
