+++ shared/src/Sienar.Utils/Data/IResultActor.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Sienar.Infrastructure;
 
-namespace Sienar.Services;
+namespace Sienar.Data;
 
 /// <summary>
 /// A service that accepts no input and returns output
