+++ shared/src/Sienar.Utils/Extensions/SienarUtilsServiceCollectionExtensions.cs
@@ -11,7 +11,6 @@ using Sienar.Infrastructure;
 using Sienar.Menus;
 using Sienar.Processors;
 using Sienar.Security;
-using Sienar.Services;
 
 namespace Sienar.Extensions;
