+++ shared/src/Sienar.Utils/Processors/IStatusProcessor.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Sienar.Infrastructure;
-using Sienar.Services;
+using Sienar.Data;
 
 namespace Sienar.Processors;
